bot/requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
fastapi==0.100.0
uvicorn[standard]==0.23.2
python-multipart==0.0.20
